+++ src/index.ts
@@ -7,15 +7,26 @@ interface Options {
   persistence?: number
 }
 
+function cerp (p0: number, p1: number, p2: number, p3: number, x: number): number {
+  return (
+    (-0.5 * p0 + 1.5 * p1 - 1.5 * p2 + 0.5 * p3) * Math.pow(x, 3) +
+    (p0 - 2.5 * p1 + 2 * p2 - 0.5 * p3) * Math.pow(x, 2) +
+    (-0.5 * p0 + 0.5 * p2) * x +
+    p1
+  )
+}
+
 function generate1DNoiseFn (source: Uint8Array): (ox: number) => number {
   const sx = source.length
   return ox => {
     const x = ox % sx + sx
     const _x = Math.floor(x)
     const xi = x - _x
+    const x0 = ((_x - 1) + sx) % sx
     const x1 = _x % sx
     const x2 = (_x + 1) % sx
-    return lerp(source[x1], source[x2], xi)
+    const x3 = (_x + 2) % sx
+    return cerp(source[x0], source[x1], source[x2], source[x3], xi)
   }
 }
 
@@ -25,11 +36,15 @@ function generate2DNoiseFn (source: Uint8Array[]): (ox: number, oy: number) => n
     const [x, y] = [ox % sx + sx, oy % sy + sy]
     const [_x, _y] = [Math.floor(x), Math.floor(y)]
     const [xi, yi] = [x - _x, y - _y]
+    const [x0, y0] = [((_x - 1) + sx) % sx, ((_y - 1) + sy) % sy]
     const [x1, y1] = [_x % sx, _y % sy]
     const [x2, y2] = [(_x + 1) % sx, (_y + 1) % sy]
-    return lerp(
-      lerp(source[x1][y1], source[x2][y1], xi),
-      lerp(source[x1][y2], source[x2][y2], xi),
+    const [x3, y3] = [(_x + 2) % sx, (_y + 2) % sy]
+    return cerp(
+      cerp(source[x0][y0], source[x1][y0], source[x2][y0], source[x3][y0], xi),
+      cerp(source[x0][y1], source[x1][y1], source[x2][y1], source[x3][y1], xi),
+      cerp(source[x0][y2], source[x1][y2], source[x2][y2], source[x3][y2], xi),
+      cerp(source[x0][y3], source[x1][y3], source[x2][y3], source[x3][y3], xi),
       yi
     )
   }
@@ -41,19 +56,41 @@ function generate3DNoiseFn (source: Uint8Array[][]): (ox: number, oy: number, oz
     const [x, y, z] = [ox % sx + sx, oy % sy + sy, oz % sz + sz]
     const [_x, _y, _z] = [Math.floor(x), Math.floor(y), Math.floor(z)]
     const [xi, yi, zi] = [x - _x, y - _y, z - _z]
+    const [x0, y0, z0] = [((_x - 1) + sx) % sx, ((_y - 1) + sy) & sy, ((_z - 1) + sz) % sz]
     const [x1, y1, z1] = [_x % sx, _y % sy, _z % sz]
     const [x2, y2, z2] = [(_x + 1) % sx, (_y + 1) % sy, (_z + 1) % sz]
-    const bottom = lerp(
-      lerp(source[x1][y1][z1], source[x2][y1][z1], xi),
-      lerp(source[x1][y2][z1], source[x2][y2][z1], xi),
-      yi
+    const [x3, y3, z3] = [(_x + 2) % sx, (_y + 2) % sy, (_z + 2) % sz]
+    return cerp(
+      cerp(
+        cerp(source[x0][y0][z0], source[x1][y0][z0], source[x2][y0][z0], source[x3][y0][z0], xi),
+        cerp(source[x0][y1][z0], source[x1][y1][z0], source[x2][y1][z0], source[x3][y1][z0], xi),
+        cerp(source[x0][y2][z0], source[x1][y2][z0], source[x2][y2][z0], source[x3][y2][z0], xi),
+        cerp(source[x0][y3][z0], source[x1][y3][z0], source[x2][y3][z0], source[x3][y3][z0], xi),
+        yi
+      ),
+      cerp(
+        cerp(source[x0][y0][z1], source[x1][y0][z1], source[x2][y0][z1], source[x3][y0][z1], xi),
+        cerp(source[x0][y1][z1], source[x1][y1][z1], source[x2][y1][z1], source[x3][y1][z1], xi),
+        cerp(source[x0][y2][z1], source[x1][y2][z1], source[x2][y2][z1], source[x3][y2][z1], xi),
+        cerp(source[x0][y3][z1], source[x1][y3][z1], source[x2][y3][z1], source[x3][y3][z1], xi),
+        yi
+      ),
+      cerp(
+        cerp(source[x0][y0][z2], source[x1][y0][z2], source[x2][y0][z2], source[x3][y0][z2], xi),
+        cerp(source[x0][y1][z2], source[x1][y1][z2], source[x2][y1][z2], source[x3][y1][z2], xi),
+        cerp(source[x0][y2][z2], source[x1][y2][z2], source[x2][y2][z2], source[x3][y2][z2], xi),
+        cerp(source[x0][y3][z2], source[x1][y3][z2], source[x2][y3][z2], source[x3][y3][z2], xi),
+        yi
+      ),
+      cerp(
+        cerp(source[x0][y0][z3], source[x1][y0][z3], source[x2][y0][z3], source[x3][y0][z3], xi),
+        cerp(source[x0][y1][z3], source[x1][y1][z3], source[x2][y1][z3], source[x3][y1][z3], xi),
+        cerp(source[x0][y2][z3], source[x1][y2][z3], source[x2][y2][z3], source[x3][y2][z3], xi),
+        cerp(source[x0][y3][z3], source[x1][y3][z3], source[x2][y3][z3], source[x3][y3][z3], xi),
+        yi
+      ),
+      zi
     )
-    const top = lerp(
-      lerp(source[x1][y1][z2], source[x2][y1][z2], xi),
-      lerp(source[x1][y2][z2], source[x2][y2][z2], xi),
-      yi
-    )
-    return lerp(bottom, top, zi)
   }
 }
 
@@ -80,7 +117,6 @@ export function generateCylinder (circumference: number, height: number, options
       const rdx = nx * TWO_PI
       const a = r * Math.sin(rdx)
       const b = r * Math.cos(rdx)
-
       return noise(a * freq, b * freq, y * freq) * (amplitude * Math.pow(persistence, octave))
     }).reduce((total, num) => total + num, 0)
     / (2 - (1 / Math.pow(2, octaves - 1)))
@@ -126,37 +162,3 @@ function processOptions (options: Options): Options {
     persistence: typeof options.persistence === 'number' ? options.persistence : 0.5
   }
 }
-
-// ---------------------------------------------------------------------------------------------------------------------
-
-const height = 512
-const width = 1024
-
-const canvas = document.querySelector('canvas')
-canvas.setAttribute('height', String(height))
-canvas.setAttribute('width', String(width))
-
-const ctx = canvas.getContext('2d')
-const imageData = ctx.createImageData(width, height)
-
-const numPoints = width * height
-const startTime = Date.now()
-
-// const noise = generateCylinder(width, height, { frequency: 0.05, octaves: 32 })
-const noise = generateLine(width, { frequency: 0.2 })
-for (let x = 0; x < width; x++) {
-  for (let y = 0; y < height; y++) {
-    const i = (x + y * width) * 4
-    const value = noise[x]
-    imageData.data[i] = value
-    imageData.data[i + 1] = value
-    imageData.data[i + 2] = value
-    imageData.data[i + 3] = 255
-  }
-}
-
-const time = Date.now() - startTime
-const rate = Math.floor(numPoints / time)
-console.log(`${time}ms (${rate}pt/ms)`)
-
-ctx.putImageData(imageData, 0, 0)